fix(vcs): address review — flush race, JSDoc, LOC decomposition
